MultiAIagent/risk_evaluation_agent/requirements.txt
bedrock-agentcore-starter-toolkit
boto3
pydantic
pydantic-core
typing-extensions
annotated-types
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-instrumentation
opentelemetry-instrumentation-threading
opentelemetry-instrumentation-logging
watchdog
requests
PyYAML
